Switch implementation to jda-reactor.

+++ src/main/kotlin/pathfinder/weatherBot/weather/events/Wildfire.kt
@@ -7,17 +7,17 @@ import pathfinder.weatherBot.weather.precipitation.rain.Thunderstorm
 open class Wildfire(private val weather: Weather) : Event<Wildfire> {
     companion object {
         operator fun invoke(weather: Weather): Wildfire? {
-            return weather.hour.day.forecast.biome.humidity.toInt().let {
+            return weather.hour.day.forecast.biome.humidity.let {
                 if (
-                        (weather.hour.temp - 90 - it - (weather.precipitation?.fireRetardance ?: 0)) / 2 <= 100
-                        || (weather.precipitation is Thunderstorm && (1 d 100) <= 60 - it)
+                    (weather.hour.temp - 90 - it - (weather.precipitation?.fireRetardance ?: 0)) / 2 <= 100
+                    || (weather.precipitation is Thunderstorm && (1 d 100) <= 60 - it)
                 ) Wildfire(weather) else null
             }
         }
     }
 
     override fun progress(weather: Weather): Wildfire? =
-        if(TODO("extinguish?")) null else Wildfire(weather)
+        if (TODO("extinguish?")) null else Wildfire(weather)
 
     override fun description(prev: Wildfire?): String = TODO()
     override val finished = TODO()
